将日志清理逻辑移动到插件初始化

+++ src/main/java/com/stonewu/sitepush/SitePushPlugin.java
@@ -10,15 +10,13 @@ import run.halo.app.plugin.BasePlugin;
 import com.stonewu.sitepush.scheme.PushLog;
 import com.stonewu.sitepush.scheme.PushUnique;
 
+import java.time.Instant;
 import java.util.HashMap;
 import java.util.List;
 
 /**
- * <p>Plugin main class to manage the lifecycle of the plugin.</p>
- * <p>This class must be public and have a public constructor.</p>
- * <p>Only one main class extending {@link BasePlugin} is allowed per plugin.</p>
- *
- * @author guqing
+ * 站点收录推送插件
+ * @author stonewu
  * @since 1.0.0
  */
 @Component
@@ -29,7 +27,8 @@ public class SitePushPlugin extends BasePlugin {
 
     private final ExtensionClient client;
 
-    public SitePushPlugin(PluginWrapper wrapper, SchemeManager schemeManager, ExtensionClient client) {
+    public SitePushPlugin(PluginWrapper wrapper, SchemeManager schemeManager,
+        ExtensionClient client) {
         super(wrapper);
         this.schemeManager = schemeManager;
         this.client = client;
@@ -41,17 +40,20 @@ public class SitePushPlugin extends BasePlugin {
         schemeManager.register(PushUnique.class);
         List<PushUnique> list = client.list(PushUnique.class, null, null);
         list.forEach(pushUnique -> {
-            log.info("init---初始化---------{}----|{}", pushUnique.getCacheKey(), pushUnique.getPushStatus());
             GlobalCache.PUSH_CACHE.put(pushUnique.getCacheKey(), pushUnique);
+        });
+        List<PushLog> logList = client.list(PushLog.class, null, null);
+        logList.forEach(pushLog -> {
+            long epochSecond = Instant.now().getEpochSecond();
+            // 每次启动插件删除一个月前的推送日志
+            if (pushLog.getCreateTime() + 60 * 24 * 30 < epochSecond) {
+                client.delete(pushLog);
+            }
         });
     }
 
     @Override
     public void stop() {
-        List<PushLog> list = client.list(PushLog.class, null, null);
-        list.forEach(pushLog -> {
-            client.delete(pushLog);
-        });
         GlobalCache.PUSH_CACHE = new HashMap<>();
         Scheme pushLogScheme = schemeManager.get(PushLog.class);
         Scheme pushUniqueScheme = schemeManager.get(PushUnique.class);
